pyproject.toml
[project]
name = "line-echo-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "gunicorn>=24.1.1",
    "line-bot-sdk>=3.22.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
